Change float type to f64

+++ src/math/mod.rs
@@ -1,7 +1,7 @@
 pub mod polynomial;
 pub mod term;
 
-pub fn sqrt(x: f32) -> f32 {
+pub fn sqrt(x: f64) -> f64 {
     if x < 0.0 {
         panic!("Cannot compute the square root of a negative number");
     }
@@ -20,15 +20,15 @@ pub fn sqrt(x: f32) -> f32 {
     guess
 }
 
-pub fn fract_or_float(x: f32) -> String {
+pub fn fract_or_float(x: f64) -> String {
     let tol = 1e-9;
     for den in 1..=100 {
-        let num = (x * den as f32).round();
-        if (x - num / den as f32).abs() < tol {
+        let num = (x * den as f64).round();
+        if (x - num / den as f64).abs() < tol {
             if den == 1 {
-                return format!("{}", num as i32);
+                return format!("{}", num as i64);
             } else {
-                return format!("{}/{}", num as i32, den);
+                return format!("{}/{}", num as i64, den);
             }
         }
     }
@@ -42,8 +42,8 @@ mod tests {
     #[test]
     fn sqrt_test() {
         let eps = 1e-5;
-        assert!((sqrt(0f32) - 0f32).abs() < eps);
-        assert!((sqrt(4f32) - 2f32).abs() < eps);
-        assert!((sqrt(9f32) - 3f32).abs() < eps);
+        assert!((sqrt(0f64) - 0f64).abs() < eps);
+        assert!((sqrt(4f64) - 2f64).abs() < eps);
+        assert!((sqrt(9f64) - 3f64).abs() < eps);
     }
 }
